do not show undeployed changesets when verbose is false
